+++ lib/src/drm/common/ENCRYPTION_SCHEME.ts
@@ -1,0 +1,14 @@
+import { CBCS } from './CBCS.js';
+import { CENC } from './CENC.js';
+
+/**
+ * Encryption schemes.
+ *
+ * @group DRM
+ *
+ * @beta
+ */
+export const ENCRYPTION_SCHEME = {
+	CENC: CENC as typeof CENC,
+	CBCS: CBCS as typeof CBCS,
+} as const;