+++ lib/src/drm/common/HW_SECURE_DECODE.ts
@@ -1,0 +1,8 @@
+/**
+ * Hardware Secure Decode Robustness.
+ *
+ * @group DRM
+ *
+ * @beta
+ */
+export const HW_SECURE_DECODE = 'HW_SECURE_DECODE';