+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to
   ([#131](https://github.com/streaming-video-technology-alliance/common-media-library/issues/131))
 - Add `prefix` and `localName` support to the XML types and utilities
   ([#145](https://github.com/streaming-video-technology-alliance/common-media-library/issues/145))
+- Common DRM Models and Constants 
+  ([#154](https://github.com/streaming-video-technology-alliance/common-media-library/issues/154))
 
 ## [0.9.0] - 2025-02-21